+++ client/src/game/tools/variants/map.ts
@@ -69,7 +69,7 @@ class MapTool extends Tool {
     }
 
     removeRect(reset = true): void {
-        if (this.shape && reset && this.active) {
+        if (this.shape && reset && this.active.value) {
             this.shape.refPoint = this.ogRP!;
             this.shape.w = this.ogW!;
             this.shape.h = this.ogH!;
@@ -101,7 +101,7 @@ class MapTool extends Tool {
         this.startPoint = startPoint;
         const layer = floorStore.currentLayer.value!;
 
-        this.active = true;
+        this.active.value = true;
 
         this.rect = new Rect(cloneP(this.startPoint), 0, 0, { fillColour: "rgba(0,0,0,0)", strokeColour: "black" });
         this.state.hasRect = true;
@@ -112,7 +112,7 @@ class MapTool extends Tool {
 
     // eslint-disable-next-line @typescript-eslint/require-await
     async onMove(lp: LocalPoint): Promise<void> {
-        if (!this.active || this.rect === undefined || this.startPoint === undefined) return;
+        if (!this.active.value || this.rect === undefined || this.startPoint === undefined) return;
 
         const endPoint = l2g(lp);
 
@@ -126,9 +126,9 @@ class MapTool extends Tool {
 
     // eslint-disable-next-line @typescript-eslint/require-await
     async onUp(): Promise<void> {
-        if (!this.active || this.rect === undefined) return;
+        if (!this.active.value || this.rect === undefined) return;
 
-        this.active = false;
+        this.active.value = false;
 
         if (selectionState.state.selection.size !== 1) {
             this.removeRect();